test: fix broken tests
- Removed tests related to force submission of jobs when version is unknown.
- Updated test cases to focus on version negotiation without relying on force submission.
- Cleaned up imports in `test_version_negotiation.py`.
- Added helper `grant_job_inbox_access` in `utils.py` for managing job inbox permissions.
- Updated `uv.lock` with latest uv
- Added `syft-rds` to CI testing
