+++ .github/workflows/unit-tests.yml
@@ -111,6 +111,29 @@ jobs:
       - name: Run job tests
         run: uv run pytest packages/syft-job/tests/ -v
 
+  rds-tests:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+
+      - name: Install the project
+        run: uv sync --all-extras
+
+      - name: Install just
+        run: |
+          curl --proto '=https' --tlsv1.2 -sSf https://just.systems/install.sh | bash -s -- --to /usr/local/bin
+
+      - name: Run RDS tests
+        run: just test-unit-rds
+
   enclave-tests:
     runs-on: ubuntu-latest
     steps:
