+++ tests/unit/test_dir_autocomplete.py
@@ -10,11 +10,17 @@ def test_dir_returns_only_public_api():
     assert "email" in public_names
     assert "sync" in public_names
     assert "peers" in public_names
-    assert "jobs" in public_names
-    assert "datasets" in public_names
     assert "add_peer" in public_names
-    assert "create_dataset" in public_names
-    assert "submit_python_job" in public_names
+    assert "approve_peer_request" in public_names
+    assert "delete_syftbox" in public_names
+
+    # Domain-free: no dataset/job surface
+    assert "jobs" not in public_names
+    assert "datasets" not in public_names
+    assert "dataset_manager" not in public_names
+    assert "create_dataset" not in public_names
+    assert "submit_python_job" not in public_names
+    assert "process_approved_jobs" not in public_names
 
     # Hides Pydantic internals
     assert "model_dump" not in public_names
